solution/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==8.0.3
cryptography==3.4.7
defusedxml==0.6.0
dlib==19.22.1
elementpath==2.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
importlib-resources==5.1.3
isodate==0.6.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.0
more-itertools==8.7.0
numpy==1.21.4
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
Pillow==8.4.0
portend==2.7.1
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tempora==4.0.2
urllib3==1.26.4
xmlschema==1.6.2
xmlsec==1.3.10
zc.lockfile==2.0
Jinja2==3.0.1
bcrypt==3.2.0
adafruit-circuitpython-fingerprint==2.2.1
ws4py==0.5.1
fingerprint-enhancer==0.0.12
Pillow==8.4.0
scikit-image==0.19.0
fingerprint-feature-extractor==0.0.7
scikit-learn==1.0.2
scipy==1.7.3
matplotlib==3.5.1
packaging==21.3
pyparsing==3.0.6
